requirements.txt
pillow
pytoshop
requests
cairosvg
six
six